refactor: 경로생성 action버튼 위젯명 변경, 색상, 자이즈 상수화
